+++ R/xStandardizedMeanDifference.R
@@ -67,3 +67,10 @@ estimateStandardizedMeanDifference <- function(...,main=NULL,digits=3) {
   results <- .formatList(list(results),main=main,digits=digits) 
   return(results)
 }
+
+### Confidence Interval Plots
+
+plotStandardizedMeanDifference <- function(...,main=NULL,digits=3,ylab="Standardized Mean Difference",xlab="",mu=0,line=NULL,rope=NULL,conf.level=.95,values=TRUE,pos=2,connect=FALSE,ylim=NULL,add=FALSE,pch=24,col="black",offset=0,intervals=TRUE) {
+  results <- estimateStandardizedMeanDifference(...,mu=mu,conf.level=conf.level,main=main,digits=digits)
+  plotIntervals(results,add=add,main=main,xlab=xlab,ylab=ylab,ylim=ylim,values=values,line=line,rope=rope,digits=digits,connect=connect,pos=pos,pch=pch,col=col,offset=offset,intervals=intervals)
+}